ref: square payments catr and orders

+++ server/models/OrderItem.js
@@ -14,7 +14,7 @@ const OrderItem = sequelize.define("OrderItem", {
     allowNull: false,
   },
   price: {
-    type: DataTypes.FLOAT,
+    type: DataTypes.DECIMAL(10, 2),
     allowNull: false,
   },
   pricingType: {
